Split sonicLiquidFluid model into different files for foam-extend, OpenFOAM ESI and OF Foundation codes (similar to pimpleFluid).
