+++ src/solids4FoamModels/fluidModels/sonicLiquidFluid/sonicLiquidFluid.H
@@ -26,165 +26,26 @@ Class
     sonicLiquidFluid
 
 Description
-    Transient solver for trans-sonic/supersonic, laminar flow of a
-    compressible liquid. Assumes a linear equation of state between
-	density and pressure related by the bulk modulus.
+    sonicLiquidFoam fluid model.
 
 Author
-    Iago Lessa de Oliveira, FEIS/UNESP/UCD.
-    Philip Cardiff, UCD. All rights reserved.
+    Based on sonicLiquidFoam.
+    Ported by Philip Cardiff and Iago Oliveira.
 
 SourceFiles
     sonicLiquidFluid.C
 
 \*---------------------------------------------------------------------------*/
 
-#ifndef sonicLiquidFluid_H
-#define sonicLiquidFluid_H
-
-#include "fluidModel.H"
-#include "pimpleControl.H"
-#include "uniformDimensionedFields.H"
+// There are significant differences between OpenFOAM/FOAM versions so to
+// improve readability and maintainability, the ported versions will be given
+// in separate files
 #ifdef OPENFOAMFOUNDATION
-    #include "pressureReference.H"
-    #include "fvcSmooth.H"
+    #include "sonicLiquidFluid.foundation.H"
 #elif OPENFOAMESI
-    #include "pressureControl.H"
-    #include "fvcSmooth.H"
-#endif
-
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-namespace Foam
-{
-
-namespace fluidModels
-{
-
-/*---------------------------------------------------------------------------*\
-                        Class sonicLiquidFluid Declaration
-\*---------------------------------------------------------------------------*/
-
-class sonicLiquidFluid
-:
-    public fluidModel
-{
-    // Private data
-
-        // Transport properties dictionary
-        IOdictionary transportProperties_;
-
-        //- Dynamic viscosity
-        dimensionedScalar mu_;
-
-		//- Thermodynamic properties
-		IOdictionary thermodynamicProperties_;
-
-		//- Ref. fluid density
-		const dimensionedScalar rho0_;
-
-		//- Ref. pressure
-		const dimensionedScalar p0_;
-
-        //- const Ref. of psi field
-        const dimensionedScalar psi_;
-
-		//- Density offset
-        const volScalarField rhoO_;
-
-        //- Fluid density
-        volScalarField rho_;
-
-#ifdef OPENFOAMESIORFOUNDATION
-        //- Creates and initialises the momentum field rhoUf
-        autoPtr<surfaceVectorField> rhoUf_;
-#endif
-        //- rAU field
-        volScalarField rAU_;
-
-        //- Cumulative continuity error
-        scalar cumulativeContErr_;
-
-    // Private Member Functions
-
-        //- Calculates and prints the continuity errors
-        void compressibleContinuityErrs();
-
-        //- Solve density equation
-        void solveRhoEqn();
-
-        //- Phi correction for mesh motion
-        void CorrectFlux();
-
-        //- Print compressible Courant number
-		void compressibleCourantNo();
-
-        //- Solve pressure equation
-		void solvePEqn
-		(
-			const fvVectorMatrix& UEqn
-		);
-
-        //- Disallow default bitwise copy construct
-        sonicLiquidFluid(const sonicLiquidFluid&);
-
-        //- Disallow default bitwise assignment
-        void operator=(const sonicLiquidFluid&);
-
-
-public:
-
-    //- Runtime type information
-    TypeName("sonicLiquidFluid");
-
-    // Constructors
-
-        //- Construct from components
-        sonicLiquidFluid
-        (
-            Time& runTime,
-            const word& region = dynamicFvMesh::defaultRegion
-        );
-
-
-    // Destructor
-
-        virtual ~sonicLiquidFluid()
-        {}
-
-
-    // Member Functions
-
-        // Access
-
-            //- Patch viscous force (N/m2)
-            virtual tmp<vectorField> patchViscousForce
-            (
-                const label patchID
-            ) const;
-
-            //- Patch pressure force (N/m2)
-            virtual tmp<scalarField> patchPressureForce
-            (
-                const label patchID
-            ) const;
-
-
-        // Edit
-
-            //- Evolve the fluid model
-            virtual bool evolve();
-};
-
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
-} // End namespace fluidModels
-} // End namespace Foam
-
-// * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * //
-
+    #include "sonicLiquidFluid.esi.H"
+#else
+    #include "sonicLiquidFluid.foamextend.H"
 #endif
 
 // ************************************************************************* //
